+++ packages/i18n/types/TranslatableString.ts
@@ -1,0 +1,11 @@
+/**
+ * A source string with an explicit translation key.
+ *
+ * Used by `src/resBundle.ts` and `$L/$L.ts`.
+ */
+interface TranslatableString {
+	key?: string;
+	value: string;
+}
+
+export type {TranslatableString};
